apps/api/pyproject.toml
[project]
name = "flood-risk-api"
version = "0.0.0"
description = "FastAPI backend skeleton for the Flood Risk project."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "psycopg[binary]>=3.2",
  "redis>=5.0",
  "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27",
  "jsonschema>=4.21",
  "mypy>=1.11",
  "pytest>=8.3",
  "pyyaml>=6",
  "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
